Adjust Game and Team enums to use custom UpperStrEnum class
enum.StrEnum was added in Python 3.11, but since this module supports 3.10 this cannot be used. Instead, we can implement our own enum class to take its place. This also means that we can change the implementation of "_generate_next_value_" to allow us to return upper case values, and therefore we can use "auto()" instead of duplicating the member names as strings.